angular2  bundle and react Bundle integration

+++ src/nodefony/bundles/frameworkBundle/services/webpackService.js
@@ -10,15 +10,12 @@ const webpackMerge = require('webpack-merge'); // used to merge webpack configs
 const ExtractTextPluginCss = require('extract-text-webpack-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 
+module.exports = nodefony.registerService("webpack", function(){
 
-nodefony.registerService("webpack", function(){
-
-
-
-	var babelRule  = function(basename){
+	var babelRule  = function(/*basename*/){
 		return {
-                    	test: /\.es6$/,
-			exclude: /node_modules/,
+      		test: new RegExp("\.es6$"),
+			exclude: new RegExp("node_modules"),
 			use: [
 				{
 					loader: 'babel-loader',
@@ -30,33 +27,33 @@ nodefony.registerService("webpack", function(){
 		};
 	};
 
-	var cssRule  = function(basename, production){
+	var cssRule  = function(/*basename, production*/){
 		return {
-			test: /\.css$/,
+			test: new RegExp("\.css$"),
 			use: ExtractTextPluginCss.extract({
 				use: 'css-loader'
 			})
 		};
 	};
 
-	var sassRule  = function(basename){
+	var sassRule  = function(/*basename*/){
 		return {
-                    test: /.scss$/,
-                    use: [
-                        {
-                            loader: 'style-loader'
-                        }, {
-                            loader: 'css-loader'
-                        }, {
-                            loader: 'sass-loader'
-                        }
-                    ]
-                };
+        test: new RegExp(".scss$"),
+        use: [
+            {
+                loader: 'style-loader'
+            }, {
+                loader: 'css-loader'
+            }, {
+                loader: 'sass-loader'
+            }
+        ]
+    };
 	};
 
-	var lessRule  = function(basename){
+	var lessRule  = function(/*basename*/){
 		return {
-			test: /\.less$/,
+			test: new RegExp("\.less$"),
 			use: ExtractTextPluginCss.extract({
 				use: [
 					//'style-loader',
@@ -78,8 +75,7 @@ nodefony.registerService("webpack", function(){
          */
 	var fontsRule = function(basename){
 		return {
-			test: /\.(eot|woff2?|svg|ttf)([\?]?.*)$/,
-			//use: 'file-loader?name=[name].[ext]&publicPath=/'+basename+'/assets/fonts/&outputPath=/assets/fonts/',
+			test: new RegExp("\.(eot|woff2?|svg|ttf)([\?]?.*)$"),
 			use: 'file-loader?name=[name].[ext]&publicPath=/'+basename+'&outputPath=/assets/fonts/',
 		};
         };
@@ -89,17 +85,17 @@ nodefony.registerService("webpack", function(){
          */
 	var imagesRule = function(basename){
 		return {
-			test: /\.(jpg|png|gif)$/,
-          		//use: 'file-loader?name=[name].[ext]&publicPath=/'+basename+'/assets/images/&outputPath=/assets/images/'
-          		use: 'file-loader?name=[name].[ext]&publicPath=/'+basename+'&outputPath=/assets/images/'
+			test: new RegExp("\.(jpg|png|gif)$"),
+  		use: 'file-loader?name=[name].[ext]&publicPath=/'+basename+'&outputPath=/assets/images/'
 		};
         };
 
 	var defaultConfig = function(name, Path){
+		let basename = null ;
 		if ( Path ){
-			var basename = path.basename(Path);
+			basename = path.basename(Path);
 		}else{
-			var basename ="assets";
+			basename ="assets";
 		}
 		var public = Path ? path.resolve( Path, "Resources", "public") : null;
 		var devtool = this.production ? false : 'source-map' ;
@@ -137,7 +133,12 @@ nodefony.registerService("webpack", function(){
 			super ("WEBPACK", container);
 			this.production = ( this.kernel.environment === "prod" ) ?  true :  false ;
 			this.defaultConfig = defaultConfig.call(this, "nodefony");
-			this.pathCache = path.resolve(this.kernel.rootDir, "tmp", "webpack") ;
+			//this.pathCache = path.resolve(this.kernel.rootDir, "tmp", "webpack") ;
+			this.pathCache = this.kernel.cacheWebpack ;
+			this.host = this.kernel.hostHttps ;
+			this.webpack = webpack ;
+			this.version = this.getWebpackVersion();
+
 			if ( this.production ){
 				try {
 					if ( ! fs.existsSync( this.pathCache ) ){
@@ -146,31 +147,38 @@ nodefony.registerService("webpack", function(){
 				}catch(e){
 					this.logger(e.message, "WARNING");
 				}
-				this.kernel.listen(this, "onTerminate", () =>{
-					var res = fs.readdirSync(this.pathCache);
-					if ( res && res.length ){
-						for ( var i = 0 ; i < res.length ; i++ ){
-							fs.rmdirSync( path.resolve (this.pathCache ,res[i] ) );
+				if ( ! this.production ){
+					this.kernel.listen(this, "onTerminate", () =>{
+						var res = fs.readdirSync(this.pathCache);
+						if ( res && res.length ){
+							for ( var i = 0 ; i < res.length ; i++ ){
+								fs.rmdirSync( path.resolve (this.pathCache ,res[i] ) );
+							}
 						}
-					}
-				});
+					});
+				}
 			}
+			this.sockjs =  this.get("sockjs") ;
 		}
 
-		loggerStat (err, stats, bundle , watcher){
+		getWebpackVersion (){
+			return process.env.WEBPACK_VERSION;
+		}
+
+		loggerStat (err, stats, bundle , file, watcher){
 			if (err){
-				throw err
+				throw err;
 			}
 			const info = stats.toJson();
 			var error = stats.hasErrors();
 			if ( error ) {
-				this.logger(info.errors ,"ERROR")
+				this.logger(info.errors ,"ERROR");
 			}else{
 				if (bundle){
 					if ( watcher ){
-						this.logger( "WATCHING BUNDLE : " + bundle,"INFO");
+						this.logger( "WATCHING BUNDLE : " + bundle +" " + file,"INFO");
 					}else{
-						this.logger( "COMPILE BUNDLE : " + bundle,"INFO");
+						this.logger( "COMPILE SUCCESS BUNDLE : " + bundle +" " + file,"INFO");
 					}
 				}
 				if ( watcher ){
@@ -180,9 +188,111 @@ nodefony.registerService("webpack", function(){
 					}), "INFO");
 				}
 				if (stats.hasWarnings()) {
-					this.logger(info.warnings ,"WARNING")
+					this.logger(info.warnings ,"WARNING");
 				}
 			}
+		}
+
+		loadConfigFile (file, bundle){
+			try {
+				if ( ! ( file instanceof nodefony.fileClass ) ){
+					file = new nodefony.fileClass(file);
+				}
+			}catch(e){
+				throw e ;
+			}
+			let Path = bundle.path ;
+			let type = bundle.settings.type;
+			let publicPath = bundle.publicPath;
+			let config = null ;
+			let basename = bundle.bundleName;
+			let watch = null ;
+			let compiler = null ;
+			try {
+				switch (type){
+					case "angular" :
+						publicPath = path.resolve("/", bundle.bundleName, "dist" );
+						shell.cd(Path);
+						config = require( file.path );
+						config.output.path = path.resolve("Resources", "public", "dist") ;
+						if ( publicPath ){
+							config.output.publicPath = publicPath+"/" ;
+						}else{
+							config.output.publicPath =  "/"+path.basename(file.dirName)+"/dist/" ;
+						}
+						if ( ! this.production ){
+							config.entry.main.unshift("webpack-dev-server/client?https://"+this.host+"/");
+						}
+					break;
+					case "react" :
+						publicPath = path.resolve("/", bundle.bundleName, "dist" );
+						shell.cd(Path);
+						process.env.PUBLIC_URL = publicPath ;
+						config = require(file.path );
+						config.output.path = path.resolve("Resources", "public", "dist") ;
+						if ( publicPath ){
+							config.output.publicPath = publicPath+"/" ;
+						}else{
+							config.output.publicPath =  "/"+path.basename(file.dirName)+"/dist/" ;
+						}
+					break;
+					default :
+						config = require( file.path );
+				}
+				config.name = file.name || basename  ;
+				compiler =  webpack( config );
+				if ( this.kernel.type === "CONSOLE" ){
+					shell.cd(this.kernel.rootDir);
+					return compiler ;
+				}
+			}catch(e){
+				shell.cd(this.kernel.rootDir);
+				throw e ;
+			}
+			// WATCH
+			if ( config.watch === undefined ){
+				watch = bundle.webpackWatch ;
+				config.watch = watch || false;
+			}else{
+				watch = config.watch ;
+			}
+			if ( this.production ){
+				watch = false ;
+			}else{
+				if ( watch && this.sockjs && compiler ){
+					this.sockjs.addCompiler(  compiler, basename);
+				}
+			}
+			try {
+				let idfile = basename+"_"+file.name ;
+				let watching = null ;
+				if ( watch ){
+					watching = compiler.watch({
+						/* watchOptions */
+					}, (err, stats) => {
+						if (! err ){
+							this.logger( "RUN WEBPACK COMPILER : "+ basename +" COMPILE ENTRY POINT : \n" +this.displayConfigTable(config) ,"DEBUG");
+						}
+						this.loggerStat(err, stats, basename, file.name, true);
+					});
+					this.kernel.listen(this ,"onTerminate", ( ) => {
+						watching.close(() => {
+							this.logger("Watching Ended  " + config.context +" : "+util.inspect(config.entry) , "INFO");
+						});
+					});
+				}else{
+					if ( (this.kernel.environment === "dev" ) && (basename in this.kernel.bundlesCore) && ( ! this.kernel.isCore ) ){
+						shell.cd(this.kernel.rootDir);
+						return compiler ;
+					}
+					this.runCompiler(compiler, idfile, basename, file.name );
+				}
+			}catch(e){
+				shell.cd(this.kernel.rootDir);
+				throw e ;
+			}
+			shell.cd(this.kernel.rootDir);
+			return compiler ;
 		}
 
 		loadConfig( config , Path ){
@@ -194,10 +304,15 @@ nodefony.registerService("webpack", function(){
 
 			if ( this.production ){
 				myConf.watch = false ;
+			}else{
+				if ( nodefony.typeOf(myConf.entry) === "array"  ){
+					myConf.entry.unshift("webpack-dev-server/client?https://"+this.host+"/");
+				}
 			}
-
+			myConf.name =  basename  ;
+			let compiler = null ;
 			try {
-				var compiler =  webpack( myConf );
+				compiler =  webpack( myConf );
 				if ( this.kernel.type === "CONSOLE" ){
 					return  compiler;
 				}
@@ -205,16 +320,24 @@ nodefony.registerService("webpack", function(){
 				throw e ;
 			}
 
+			let sokjsCompiler = null ;
+			if ( ! this.production && this.sockjs && compiler ) {
+				sokjsCompiler = this.sockjs.addCompiler( compiler, basename );
+			}
+
 			if ( ! ( basename  in this.kernel.bundlesCore ) ){
 				this.logger( "WEBPACK BUNDLE : " +  basename +" WATCHING : "+ myConf.watch );
 			}
 
+			let watching = null ;
 			if ( myConf.watch ){
-				this.logger( "WEBPACK BUNDLE : "+ basename +" WATCHING ENTRY POINT : \n" + util.inspect(myConf.entry) , "DEBUG" );
-				var watching = compiler.watch({
+				watching = compiler.watch({
 					/* watchOptions */
 				}, (err, stats) => {
-					this.loggerStat(err, stats, basename, true);
+					if (! err ){
+						this.logger( "WEBPACK BUNDLE : "+ basename +" WATCHING ENTRY POINT : \n" + util.inspect(myConf.entry) , "DEBUG" );
+					}
+					this.loggerStat(err, stats, basename, basename,true);
 				});
 				this.kernel.listen(this ,"onTerminate", ( ) => {
 					watching.close(() => {
@@ -222,19 +345,18 @@ nodefony.registerService("webpack", function(){
 					});
 				});
 			}else{
-				if ( (this.kernel.environment === "dev" ) && (basename in this.kernel.bundlesCore) && ( ! this.kernel.settings.system.core ) ){
+				if ( (this.kernel.environment === "dev" ) && (basename in this.kernel.bundlesCore) && ( ! this.kernel.isCore ) ){
 					return compiler ;
 				}
-				this.logger( "WEBPACK BUNDLE : "+ basename +" COMPILE ENTRY POINT : \n" + util.inspect(myConf.entry)  );
-				this.runCompiler(compiler, basename);
+				this.runCompiler(compiler, basename, basename);
 			}
 			return compiler ;
 		}
 
-		runCompiler (compiler, bundle){
+		runCompiler (compiler, id, bundle, file){
 			try {
 				if ( this.production ){
-					var pathCache = path.resolve( this.pathCache, bundle );
+					var pathCache = path.resolve( this.pathCache, id );
 					if ( fs.existsSync( pathCache ) ){
 						return ;
 					}
@@ -244,11 +366,46 @@ nodefony.registerService("webpack", function(){
 
 					}
 				}
-				return compiler.run( (err, stats) => {
-					this.loggerStat(err, stats,  bundle);
+				return new Promise ( (resolve/*, reject*/) => {
+					this.logger( "RUN WEBPACK COMPILER : "+ file +" COMPILE ENTRY POINT : \n" +this.displayConfigTable(compiler.options) ,"DEBUG");
+					compiler.run( (err, stats) => {
+						this.loggerStat(err, stats, bundle, file);
+						return resolve(err, stats);
+					});
 				});
 			}catch(e){
 				throw e ;
+			}
+		}
+
+		displayConfigTable (config){
+			let options = {
+				head:[
+					"ENTRY NAME [name]",
+					"ENTRY",
+					"OUTPUT FILE NAME",
+					"OUTPUT LIBRARY NANE",
+					"OUTPUT TARGET",
+					"PUBLIC PATH",
+					"WATCHER"
+				]
+			} ;
+			let table = this.kernel.cli.displayTable(null, options);
+			try {
+				for ( let ele in config.entry ){
+					table.push([
+						ele,
+						config.entry[ele].toString(),
+						config.output.filename ,
+						config.output.library ,
+						config.output.libraryTarget,
+						config.output.path,
+						config.watch || ""
+					]);
+				}
+				return table.toString() ;
+			}catch(e){
+				this.logger(e, "ERROR");
 			}
 		}
 
@@ -271,6 +428,11 @@ nodefony.registerService("webpack", function(){
 			}catch(e){
 				throw e;
 			}
+		}
+
+		getProgressPlugin (handler){
+			//function handler(percentage, msg) {/* ... */}
+			return new webpack.ProgressPlugin(handler);
 		}
 	}
